refactor struct BlobWorkerData and struct GranuleMetadata
